requirements.txt
python-dotenv~=0.21.0
bs4~=0.0.1
beautifulsoup4~=4.11.1
requests~=2.28.1
pandas~=1.4.4